domiot_ecabinet/requirements.txt
fastapi[standard]
uvicorn
paho-mqtt
requests